Changed "Calibration.m" to use definition file.
Modified "Calibration.m" so that it stays the same from run to run, and the only thing that changes is the definition file.  I made a sample definition file ("sample_def.m") in the folder, which shows how to tell "Calibration.m" which data to load and initial guesses for the optimization.  I tried the optimization and fitting with the calibrations from October 10, and they work perfectly.