Made CosinePad; modified Retrieve Vels_Freqs.m to retrieve significant transients.
Made CosinePad.m.  This will pad a given transient to the given length.  The padding before the transient is zero padding, and after the transient is a cosine window of the transient.

Retrieve_Vels_Freqs.m now has functionality to retrieve significant transients, pad, and window them.  It does not yet perform frequency analysis.