Initial commit of head movement visualization scripts.
This is the first commit of the head movement visualization scripts.  In
this commit, we have "Head_Movement_Vis.m".  This creates a smiley head
and rotates it based on the data recorded in a given data file (in this
commit, Alex walking down stairs).  It relies on makeRotationMatrix.m,
which creates a rotation matrix for a given angle and axis.