Plots correlations; include more data.
Initial commit of PlotCorrelationDelay.m, which shows the correlation
between roll and pitch as a function of delay.  At this point, it looks
like there is a correlation (sometimes up to 0.75), but not for all
subjects, and more investigation is required.   Possibly younger
subjects have higher correlations.

In addition, processed two more subjects--YEI009 and YEI010.